Remove more dependency on CryptoNoteCore headers
