+++ src/CMakeLists.txt
@@ -141,7 +141,7 @@ target_link_libraries(Errors Crypto SubWallets)
 target_link_libraries(Logging Common)
 target_link_libraries(miner CryptoNoteCore Rpc System Http Crypto Errors Utilities)
 target_link_libraries(Nigel Errors)
-target_link_libraries(P2P CryptoNoteCore upnpc-static)
+target_link_libraries(P2P upnpc-static)
 target_link_libraries(Rpc P2P Utilities)
 target_link_libraries(SubWallets Common Logger)
 target_link_libraries(Utilities Common Errors)
