+++ src/CryptoNoteCore/TransactionPrefixImpl.cpp
@@ -91,7 +91,7 @@ Hash TransactionPrefixImpl::getTransactionPrefixHash() const {
 }
 
 PublicKey TransactionPrefixImpl::getTransactionPublicKey() const {
-  Crypto::PublicKey pk(NULL_PUBLIC_KEY);
+  Crypto::PublicKey pk(Constants::NULL_PUBLIC_KEY);
   m_extra.getPublicKey(pk);
   return pk;
 }
