+++ src/P2p/P2pProtocolDefinitions.h
@@ -11,7 +11,6 @@
 
 #include "crypto/crypto.h"
 #include <config/CryptoNoteConfig.h>
-#include "CryptoNoteCore/CoreStatistics.h"
 
 // new serialization
 #include "Serialization/ISerializer.h"
@@ -172,121 +171,4 @@ namespace CryptoNote
       }
     };
   };
-
-
-#ifdef ALLOW_DEBUG_COMMANDS
-  //These commands are considered as insecure, and made in debug purposes for a limited lifetime.
-  //Anyone who feel unsafe with this commands can disable the ALLOW_GET_STAT_COMMAND macro.
-
-  struct proof_of_trust
-  {
-    uint64_t peer_id;
-    uint64_t    time;
-    Crypto::Signature sign;
-
-    void serialize(ISerializer& s) {
-      KV_MEMBER(peer_id)
-      KV_MEMBER(time)
-      KV_MEMBER(sign)
-    }
-  };
-
-  inline Crypto::Hash get_proof_of_trust_hash(const proof_of_trust& pot) {
-    std::string s;
-    s.append(reinterpret_cast<const char*>(&pot.peer_id), sizeof(pot.peer_id));
-    s.append(reinterpret_cast<const char*>(&pot.time), sizeof(pot.time));
-    return Crypto::cn_fast_hash(s.data(), s.size());
-  }
-
-  struct COMMAND_REQUEST_STAT_INFO
-  {
-    enum { ID = P2P_COMMANDS_POOL_BASE + 4 };
-
-    struct request
-    {
-      proof_of_trust tr;
-
-      void serialize(ISerializer& s) {
-        KV_MEMBER(tr)
-      }
-    };
-
-    struct response
-    {
-      std::string version;
-      std::string os_version;
-      uint64_t connections_count;
-      uint64_t incoming_connections_count;
-      CoreStatistics payload_info;
-
-      void serialize(ISerializer& s) {
-        KV_MEMBER(version)
-        KV_MEMBER(os_version)
-        KV_MEMBER(connections_count)
-        KV_MEMBER(incoming_connections_count)
-        KV_MEMBER(payload_info)
-      }
-    };
-  };
-
-
-  /************************************************************************/
-  /*                                                                      */
-  /************************************************************************/
-  struct COMMAND_REQUEST_NETWORK_STATE
-  {
-    enum { ID = P2P_COMMANDS_POOL_BASE + 5 };
-
-    struct request
-    {
-      proof_of_trust tr;
-
-      void serialize(ISerializer& s) {
-        KV_MEMBER(tr)
-      }
-    };
-
-    struct response
-    {
-      std::list<PeerlistEntry> local_peerlist_white;
-      std::list<PeerlistEntry> local_peerlist_gray;
-      std::list<connection_entry> connections_list;
-      uint64_t my_id;
-      uint64_t local_time;
-
-      void serialize(ISerializer& s) {
-        serializeAsBinary(local_peerlist_white, "local_peerlist_white", s);
-        serializeAsBinary(local_peerlist_gray, "local_peerlist_gray", s);
-        serializeAsBinary(connections_list, "connections_list", s);
-        KV_MEMBER(my_id)
-        KV_MEMBER(local_time)
-      }
-    };
-  };
-
-  /************************************************************************/
-  /*                                                                      */
-  /************************************************************************/
-  struct COMMAND_REQUEST_PEER_ID
-  {
-    enum { ID = P2P_COMMANDS_POOL_BASE + 6 };
-
-    struct request
-    {
-      void serialize(ISerializer& s) {}
-    };
-
-    struct response
-    {
-      uint64_t my_id;
-
-      void serialize(ISerializer& s) {
-        KV_MEMBER(my_id)
-      }
-    };
-  };
-
-#endif
-
-
 }
