+++ src/Serialization/CryptoNoteSerialization.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include <CryptoNoteCore/CryptoNoteBasic.h>
+#include <CryptoNote.h>
 
 #include "crypto/chacha8.h"
 #include "Serialization/ISerializer.h"
@@ -37,6 +37,23 @@ bool serialize(EllipticCurvePoint& ecPoint, Common::StringView name, CryptoNote:
 }
 
 namespace CryptoNote {
+
+struct ParentBlockSerializer {
+    ParentBlockSerializer(ParentBlock& parentBlock, uint64_t& timestamp, uint32_t& nonce, bool hashingSerialization, bool headerOnly) :
+      m_parentBlock(parentBlock), m_timestamp(timestamp), m_nonce(nonce), m_hashingSerialization(hashingSerialization), m_headerOnly(headerOnly) {
+    }
+
+    ParentBlock& m_parentBlock;
+    uint64_t& m_timestamp;
+    uint32_t& m_nonce;
+    bool m_hashingSerialization;
+    bool m_headerOnly;
+};
+
+inline ParentBlockSerializer makeParentBlockSerializer(const BlockTemplate& b, bool hashingSerialization, bool headerOnly) {
+    BlockTemplate& blockRef = const_cast<BlockTemplate&>(b);
+    return ParentBlockSerializer(blockRef.parentBlock, blockRef.timestamp, blockRef.nonce, hashingSerialization, headerOnly);
+}
 
 struct AccountKeys;
 struct TransactionExtraMergeMiningTag;
