+++ src/Serialization/SerializationTools.h
@@ -214,5 +214,4 @@ bool fromBinaryArray(T& object, const std::vector<uint8_t>& binaryArray) {
 
   return true;
 }
-
 }
